backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
django==5.1.1
django-allauth==65.0.1
django-cors-headers==4.4.0
django-factory-boy==1.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
factory-boy==3.3.1
faker==29.0.0
fido2==1.1.3
idna==3.10
oauthlib==3.2.2
pycparser==2.22
pyjwt==2.9.0
python-dateutil==2.9.0.post0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.1
typing-extensions==4.12.2
urllib3==2.2.3
